libgit2: Update checkout to respect context
Update libgit2-git2go checkout with a clone wrapper to respect the
context passed to it. go-git accepts context passed to it during cloning
but the git2go does not accept the context passed to it. This makes the
GitRepository.Spec.Timeout ineffective when using libgit2.

This change updates all the different libgit2 checkouts to use the new
clone function.

This is also needed to shorten the wait time for tests that depend on
failure behavior and should fail early. TestCheckout_ED25519 test would
take 30s without a context with timeout. This helps shorten the time
for such tests.

Signed-off-by: Sunny <darkowlzz@protonmail.com>