Add support for using an OCI image as source
This change defines an OCIRepository CRD that allows a user to specify a
given image to use as a source.  The contents of the image (including
images with multiple layers) are converted into a TAR and exposed to
consumers following the same conventions as the other source types.

  apiVersion: source.toolkit.fluxcd.io/v1beta1
  kind: OCIRepository
  metadata:
    name: podinfo-deploy
    namespace: flux-system
  spec:
    interval: 10m
    url: ghcr.io/stefanprodan/podinfo-deploy
    ref:
      # one of
      tag: "latest"
      digest: "sha256:45b23dee08af5e43a7fea6c4cf9c25ccf269ee113168c19722f87876677c5cb2"
      semver: "1.x"
    auth:
      # one of
      secretRef:
        name: regcred
      serviceAccountName: reg

In addition to the url, interval, timeout, ignore, and suspend keys
(all of which behave consistently with the existing source types) this
CRD also defines authentication via both an image pull Secret and
a service serviceAccountName which provide ways to contribute registry
connection credentials for the specified image.

This change also adds a new way to write to the storage archive by
streaming data from an incoming TAR without writing it to the
filesystem.  A couple of code and test functions were extracted to
reuse common functionality for both archive strategies.

Signed-off-by: Ben Hale <bhale@vmware.com>