gitrepo: Use summarize helper for success events
When source fetch failure happens, a warning event is emitted but
when resolved, there's no resolved event emitted.

This change introduces a new ResultProcessor, NotifySuccess, for
SummarizeAndPatch helper which consolidates all the messages to be
emitted as a single event. It checks for any recovery from failure
condition and emits event about the recovery. In case there's a failure
recovery and a new artifact, success event message only mentions the new
artifact success message. Recovery is implied. This helps avoid mixing
of the event types to have K8s native events friendly events that work
with event counters.

The gitrepo reconciler is modified to use the new NotifySuccess
ResultProcessor. A change in artifact is checked in reconcile() and a
Notification object is returned, populated with the success information.

When the Notification is empty/zero, and no failure recovery, no event
is emitted. Failure recovery without any Notification results in an
event with only the recovery information.

Introduce new condition "StorageOperationFailed" for use with any
failures due to storage operation. With this, all the failures are
associated with some condition, which is used to detect a failure
recovery and reported as notification.

Signed-off-by: Sunny <darkowlzz@protonmail.com>