controllers: store HelmChart Artifact with suffix
This adds a Unix suffix to the HelmChart Artifact filename, to ensure
it is unique for sequential builds triggered due to e.g. a controller
restart.

The result of this is that consumers who _think_ they are fetching an
Artifact with a certain checksum run into a 404 when attempting to
download a previously advertised but now unavailable file, instead of
running into a checksum validation error (due to non-repetitive Helm
builds).

For more information, see:
https://github.com/fluxcd/source-controller/issues/610

Signed-off-by: Hidde Beydals <hello@hidde.co>