Recreate GitRepository artifacts if .spec.ignore changes
This change is intended to clarify the behavior of the source-controller
when the `.spec.ignore` field changes. Before introduction of this
change the source artifact would not to reflect changes in `.spec.ignore` until the
controller was restarted or the source revision changed. This commit ensures
that the artifact will be recreated as required based on changes to `.spec.ignore`.

Signed-off-by: Piaras Hoban <phoban01@gmail.com>