helm: optimise index loading using JSON Loading YAML files can be particularly inefficient when compared to other formats. This introduces an optimised index in JSON format, that is already sorted, therefore subsequent loads incur a loss cost in terms of time and resources.
Testing using the bitnami index (~15mb), subsequent loads were about 10 times
faster when using JSON for format for the optimised index, and 20 times faster
when loading binary.

JSON was chosen over binary given the challenges that may rise from serving
the optimised indexes to source-controller users, which is not a primary use.

Signed-off-by: Paulo Gomes <paulo.gomes@weave.works>