git clone optimization when artifact in storage
Update the GitRepositoryReconciler to attempt git clone optimization
only when an artifact is already present in storage and do not skip
reconciliation. When a partial commit is received, due to no-op clone,
copy the existing artifact to the source build directory and populate
all the associated fields in reconcileSource(). This ensure that the
reconciliation is not skipped and allows other subreconcilers to operate
on other parts of the GitRepo object like include, ignore, etc, when
attributes related to them change but the remote repository has not
changed.

Remove git.NoChangeError as no-op clone now returns a partial commit and
no error.

Signed-off-by: Sunny <darkowlzz@protonmail.com>